Add TOP_100 and TOP_1000 commands. (#50)
These commands are interesting to me for two reasons:
 - It's common to retrieve more hits than the size of a page in the first stage
   in order to feed more data into rerankers.
 - WAND and variants are reputed to work better with low `k` values while
   MAXSCORE and variants are reputed to work better with higher `k` values, so
   it would be interesting to see if this holds in this benchmark.

I took the freedom to change the default commands from `TOP_10` /
`TOP_10_COUNT` to `TOP_100` / `TOP_100_COUNT` in the `Makefile` because it felt
more like a typical `k` value to me (Lucene changed to this value for its
nightly benchmarks ~1 year ago). Happy to revert this bit if it's
controversial.